requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu128
openai-whisper
torch>=2.0.0
sounddevice
numpy
pyautogui
scipy
keyboard
requests
